Fix typo in keydown event
Thys typo was preventing the removal of the default keypress event, surprisingly
causing only some bugs when pressing SHIFT+8, SHIFT+6, SHIFT+2 and SHIFT+1 in
Google Chrome, because the keypress event treats those keycodes as DOWN, UP,
PAGEDOWN and PAGEUP, causing the menu (when is displayed) to select one of the
entries instead of actually writing the character.

This was problematic for example on our wiki, where the SHIFT+8 is the opening
parentheses, causing this weird behavior on Google Chrome when people were
typing that character for autocompletion (we have quite a lot of pages with
such characters because of disambiguations).

This stupid typo has made me lost 3 hours of debugging the internals of
jQuery-ui and jQuery itself without noticing the typo until I realized the name
inspecting some very internal jQuery functions on the unbind function! </rant>

Change-Id: I01604fc2b7255a589191a069f323254eb48d8d18