Refactor to use the core prefixsearch API module
LinkSuggest has maintained its own API for fetching suggestions for a long while, which has become more and more unnecessary in recent years. MediaWiki 1.23 introduced the prefixsearch API module, which is more than sufficient for providing suggestions for links, and supports resolving redirect targets. That same core API module is already used extensively in MediaWiki across both WikiEditor and VisualEditor as part of the "Insert link" features that those extensions have.

This is a breaking change for any external consumers of the linksuggest API module, since this change removes that module entirely. For MediaWiki consumers that do not manually make use of the linksuggest API (e.g those who just install the extension and go - which are likely the majority of people), there should be no negative behaviour from this change.

Bug: T273898
Change-Id: I29035c2253e033122abe39405942d260e251e619